Fix listener options on Robot.catchAll
Unlike other Listener-creation methods, Robot.catchAll cannot directly
rely on the Listener constructor to handle the optional options argument
(because it wants to wrap the callback). Instead of duplicating the
argument shuffling logic, we create a listener the "normal" way (just
pass through the provided arguments) and then wrap the final callback
with the custom catchAll logic.